Package fixes and more clean up
