+++ ui/common/src/commonMain/kotlin/social/androiddev/common/decompose/CoroutineScopeExt.kt
@@ -1,0 +1,13 @@
+package social.androiddev.common.decompose
+
+import com.arkivanov.essenty.lifecycle.Lifecycle
+import com.arkivanov.essenty.lifecycle.doOnDestroy
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.cancel
+import kotlin.coroutines.CoroutineContext
+
+fun CoroutineScope(context: CoroutineContext, lifecycle: Lifecycle): CoroutineScope {
+    val scope = CoroutineScope(context)
+    lifecycle.doOnDestroy(scope::cancel)
+    return scope
+}
