+++ ui/common/src/commonMain/kotlin/social/androiddev/common/decompose/LifecycleOwnerExt.kt
@@ -1,0 +1,8 @@
+package social.androiddev.common.decompose
+
+import com.arkivanov.essenty.lifecycle.LifecycleOwner
+import kotlinx.coroutines.CoroutineScope
+import kotlin.coroutines.CoroutineContext
+
+fun LifecycleOwner.coroutineScope(context: CoroutineContext): CoroutineScope =
+    CoroutineScope(context, lifecycle)
