+++ ui/root/src/commonMain/kotlin/social/androiddev/root/composables/RootContent.kt
@@ -23,7 +23,7 @@ import social.androiddev.root.navigation.SplashComponent
 import social.androiddev.signedin.composables.SignedInRootContent
 import social.androiddev.signedin.navigation.SignedInRootComponent
 import social.androiddev.signedout.composables.SignedOutRootContent
-import social.androiddev.signedout.navigation.SignedOutRootComponent
+import social.androiddev.signedout.root.SignedOutRootComponent
 
 /**
  * App root composable that delegates business logic
