+++ ui/signed-out/src/commonMain/kotlin/social/androiddev/signedout/signin/SignInContent.kt
@@ -28,6 +28,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -36,14 +37,14 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.delay
-import social.androiddev.signedout.navigation.SignInComponent
 
 @Composable
 fun SignInContent(
-    server: String,
     component: SignInComponent,
     modifier: Modifier = Modifier,
 ) {
+
+    val state by component.state.collectAsState()
 
     Column(
         modifier = modifier
@@ -96,7 +97,7 @@ fun SignInContent(
             }
         }
         SignInWebView(
-            server = server,
+            server = state.server,
             modifier = Modifier.fillMaxSize(),
             onSignedIn = { component.onSignInSucceed() },
             onFailed = { errorVisibilityState = true }
